# Development/Thera_app/api/openapi.yaml
openapi: '3.0.3'

info:
  title: CSE186 Assignment 8 Backend
  version: 0.1.0
  description: CSE186 Assignment 8 Backend

servers:
  - url: http://localhost:3010/v0

paths:
  /login:
    post:
      description: Login
      requestBody:
        description: Request Body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        200:
          description: Login Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'

  /home:
    get:
      security:
        - bearerAuth: []
      description: Home information for user
      responses:
        401:
          description: Unauthorized
        200:
          description: Fingers Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Finger'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'

  /home/{doctor_id}:
    get:
      security:
        - bearerAuth: []
      description: Get list of patients assigned to a specific doctor
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    email:
                      type: string
                    name:
                      type: string
        403:
          description: Forbidden
        401:
          description: Unauthorized

  /getmessages/{patient_id}:
    get:
      security:
        - bearerAuth: []
      description: Get messages between the logged-in user and the specified patient
      parameters:
        - name: patient_id
          in: path
          required: true
          description: UUID of the patient
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /sendmessage/{patient_id}:
    post:
      security:
        - bearerAuth: []
      description: Send a message between the logged-in user and the specified patient
      parameters:
        - name: patient_id
          in: path
          required: true
          description: UUID of the patient
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMessage'
      responses:
        201:
          description: Message successfully sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  sent_at:
                    type: string
                    format: date-time
        400:
          description: Invalid request
        401:
          description: Unauthorized
        403:
          description: Forbidden
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'

  /doctor/{patient_id}:
    get:
      security:
        - bearerAuth: []
      description: Get the doctor assigned to a given patient
      parameters:
        - name: patient_id
          in: path
          required: true
          description: UUID of the patient
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Assigned doctor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: Doctor not found
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /create/{doctor_id}:
    post:
      summary: Doctor creates a new patient
      description: Only the logged-in doctor can create and assign a new patient
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: string
          description: ID of the logged-in doctor creating the patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  name:
                    type: string
        '400':
          description: Missing fields
        '403':
          description: Unauthorized or not the logged-in doctor
        '409':
          description: Email already in use
        '500':
          description: Internal server error
  /deletemessage/{message_id}:
    delete:
      security:
        - bearerAuth: []
      description: Delete a message by its ID (only sender can delete)
      parameters:
        - name: message_id
          in: path
          required: true
          description: UUID of the message to delete
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Message successfully deleted
        401:
          description: Unauthorized
        403:
          description: Forbidden not the sender of the message
        404:
          description: Message not found
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'

  /createdoctor/{admin_id}:
    post:
      summary: Admin creates a new doctor
      description: Only the logged-in admin can create and assign a new doctor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: admin_id
          required: true
          schema:
            type: string
          description: ID of the logged-in admin creating the doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Doctor created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  name:
                    type: string
        '400':
          description: Missing fields
        '403':
          description: Unauthorized or not the logged-in admin
        '409':
          description: Email already in use
        '500':
          description: Internal server error
  /doctors:
    get:
      summary: Get all doctors (admin only)
      description: Returns a list of all registered doctors. Only accessible by admins.
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Doctor'
        401:
          description: Unauthorized
        403:
          description: Forbidden – Only admins can access this
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /deletepatient/{patient_id}:
    delete:
      security:
        - bearerAuth: []
      description: Delete a patient and all messages involving them (as sender or recipient)
      parameters:
        - name: patient_id
          in: path
          required: true
          description: UUID of the patient to delete
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Patient and associated messages deleted
        401:
          description: Unauthorized
        403:
          description: Forbidden – Only the patient's doctor or an admin may delete
        404:
          description: Patient not found
        500:
          description: Internal server error
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /deletedoctor/{doctor_id}:
    delete:
      security:
        - bearerAuth: []
      description: >
        Deletes a doctor by ID. Also deletes all patients assigned to that doctor, 
        and all messages involving those patients.
      parameters:
        - name: doctor_id
          in: path
          required: true
          description: UUID of the doctor to delete
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Doctor, their patients, and associated messages successfully deleted
        401:
          description: Unauthorized
        403:
          description: Forbidden – Only admins may delete doctors
        404:
          description: Doctor not found
        500:
          description: Internal server error
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
   
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UnexpectedError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string

    Credentials:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    User:
      type: object
      required:
        - id
        - accessToken
        - role
        - email
        - name
      properties:
        id:
          type: string
          format: uuid
        accessToken:
          type: string
        role:
          type: string
        email:
          type: string
        name:
          type: string

    Finger:
      type: object
      required:
        - id
        - patient_id
        - data
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        data:
          type: object
          required:
            - name
            - position
          properties:
            name:
              type: string
            position:
              type: array
              items:
                type: number
              minItems: 9
              maxItems: 9

    NewMessage:
      type: object
      required:
        - content
      properties:
        content:
          type: string

    Message:
      type: object
      required:
        - id
        - sender_id
        - recipient_id
        - content
        - sent_at
      properties:
        id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        recipient_id:
          type: string
          format: uuid
        content:
          type: string
        sent_at:
          type: string
          format: date-time

    Doctor:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string